Fix hiding of anchor under navbar
Right now if you go to a link like:

https://networkservicemesh.io/events/nsmcon2019/#sponsors

The actual anchor is hidden under the navbar.

This fixes that issue for most anchors, but not the anchors to our
NSMCon talks because the anchors are enclosed in an element of type
flex, and so are not rendered as block elements.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>